+++ src/test/java/uk/ac/ucl/cs/cmic/giftcloud/restserver/FakeHttpConnectionWrapper.java
@@ -97,7 +97,7 @@ class FakeHttpConnectionWrapper implements HttpConnection {
     }
 
     @Override
-    public void setFixedLengthStreamingMode(int contentLength) {
+    public void setFixedLengthStreamingMode(long contentLength) {
 
     }
